refactor: #86etkgdqu split live-session by provider into openai/ and gemini/ sub-modules
Server: provider-specific config, types, utils, and ephemeral-token
functions move under `live_session/{openai,gemini}/`. The parent
`functions.ts` keeps the shared create/update functions and aggregates the
provider modules into the exported function list. `types.ts` becomes a
barrel re-exporting both providers' types so the frontend's existing
`~/types/live-session.type` import keeps working.

Frontend: `components/liveSession/StartNew.vue` is split into
`liveSession/gemini/StartNew.vue` (current behavior) and
`liveSession/openai/StartNew.vue`. `StartLiveSessionForm` accepts a
`voiceOptions` prop (defaulting to Gemini voices) so each variant can pass
its own provider-specific voice list. `pages/sessions/new.vue` defaults to
the Gemini variant; switching to OpenAI is a single component swap.
