+++ server/src/modules/live_session/openai/config.ts
@@ -1,4 +1,2 @@
 export const LIVE_SESSION_MODEL = "gpt-realtime-mini";
 export const LIVE_SESSION_TRANSCRIPTION_MODEL = "gpt-4o-mini-transcribe";
-
-export const GEMINI_LIVE_SESSION_MODEL = "gemini-3.1-flash-live-preview";
